+++ CHANGELOG.md
@@ -25,6 +25,7 @@
 - data-migrations/README.md with instructions on running migrations in the Docker container
 - tokenService now supports refresh tokens and signout
 - updated express middelware to fetch the access token from the cookies instead of the headers
+- removed old oauth-server package which had security vulnerabilities
 
 ## v0.0.1
 Initial Apollo Server build